post_automation/requirements.txt
flask
flask-cors
gunicorn
requests
python-dotenv
boto3
google-cloud-storage